fix hardware checks in driver files
The earlier version only defined certain functions if the prerequisite hardware had been initialized. This would cause problems if the function was called and the hardware was not initialized, as there would be nothing in the library to point to. The new version creates a function in each requisite file to check hardware, and then runs that in each other function before carrying out the function's tasks. If the hardware is initialized, it runs the function normally. If not, it logs that hardware was accessed without being initialized.